Refactored tests for datacenters
